refactored file-structure
