+++ src/components/browser-history/browser-history.tsx
@@ -1,6 +1,6 @@
 import {useState, useLayoutEffect} from 'react';
 import {Router} from 'react-router-dom';
-import { HistoryRouterProps } from '../types';
+import { HistoryRouterProps } from '../../types';
 
 function HistoryRouter({
   basename,
