+++ src/components/comment-section/comment-section.tsx
@@ -1,11 +1,11 @@
-import { ratingTypes } from '../const';
+import { ratingTypes } from '../../const';
 import { ChangeEvent, FormEvent, MouseEvent } from 'react';
 import { useParams } from 'react-router-dom';
-import useReviewChanges from '../custom-hooks/use-review-changes';
+import useReviewChanges from '../../custom-hooks/use-review-changes';
 
-import usePostingComments from '../custom-hooks/use-posting-comments';
-import { ReviewData } from '../types';
-import StarMemo from './star';
+import usePostingComments from '../../custom-hooks/use-posting-comments';
+import { ReviewData } from '../../types';
+import StarMemo from '../star/star';
 
 type ReviewChangesHookData = [
   (event: ChangeEvent<HTMLTextAreaElement>) => void,
