+++ src/components/favorite-item/favorite-item.tsx
@@ -1,34 +1,29 @@
-import { Offer } from '../types';
+import { Offer } from '../../types';
 import { Link } from 'react-router-dom';
-import { calculateCardRating } from '../utils';
+import { calculateCardRating } from '../../utils';
 import classNames from 'classnames';
 
-function PlaceCard(cardProps: Offer): JSX.Element {
-  const {isPremium, price, isFavorite, rating, id, title, type, handleHover, handleLeave, previewImage} = cardProps;
+function FavoriteItem(favoriteCard: Offer): JSX.Element {
+  const {price, isFavorite, images, rating, title, type, id} = favoriteCard;
 
-  return (
-    <article
-      className="cities__card place-card"
-      onMouseEnter={handleHover}
-      onMouseLeave={handleLeave}
-    >
-      {isPremium && <div className="place-card__mark"><span>Premium</span></div>}
-      <div className="cities__image-wrapper place-card__image-wrapper">
+  return(
+    <article className="favorites__card place-card">
+      <div className="favorites__image-wrapper place-card__image-wrapper">
         <Link to={`/offer/${id}`}>
-          <img className="place-card__image" src={previewImage} width="260" height="200" alt="Place image" />
+          <img className="place-card__image" src={images[0]} width="150" height="110" alt="Place image" />
         </Link>
       </div>
-      <div className="place-card__info">
+      <div className="favorites__card-info place-card__info">
         <div className="place-card__price-wrapper">
           <div className="place-card__price">
             <b className="place-card__price-value">&euro;{price}</b>
             <span className="place-card__price-text">&#47;&nbsp;night</span>
           </div>
-          <button className={classNames('place-card__bookmark-button', 'button', {'place-card__bookmark-button--active':isFavorite})} type="button">
+          <button className={classNames('place-card__bookmark-button', 'button', {'place-card__bookmark-button--active': isFavorite})} type="button">
             <svg className="place-card__bookmark-icon" width="18" height="19">
               <use xlinkHref="#icon-bookmark"></use>
             </svg>
-            <span className="visually-hidden">To bookmarks</span>
+            <span className="visually-hidden">In bookmarks</span>
           </button>
         </div>
         <div className="place-card__rating rating">
@@ -46,4 +41,4 @@ function PlaceCard(cardProps: Offer): JSX.Element {
   );
 }
 
-export default PlaceCard;
+export default FavoriteItem;
