+++ src/components/logo/logo.tsx
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom';
-import { AppRoute } from './app';
+import { AppRoute } from '../app/app';
 
 function Logo():JSX.Element {
   return(
