+++ src/components/place-sorting/places-sorting.tsx
@@ -1,8 +1,8 @@
-import { sortTypes } from '../const';
+import { sortTypes } from '../../const';
 import classNames from 'classnames';
-import { useAppSelector } from '../redux-hooks';
+import { useAppSelector } from '../../redux-hooks';
 import React from 'react';
-import useFiltering from '../custom-hooks/use-filtering';
+import useFiltering from '../../custom-hooks/use-filtering';
 function PLacesSorting(): JSX.Element {
 
   const [filter, isOpened, handleClick, toggleList] = useFiltering();
