+++ src/components/private-route/private-route.tsx
@@ -1,7 +1,7 @@
-import { AuthorizationStatus} from '../const';
-import { AppRoute } from './app';
+import { AuthorizationStatus} from '../../const';
+import { AppRoute } from '../app/app';
 import { Navigate } from 'react-router-dom';
-import { PrivateRouteProps } from '../types';
+import { PrivateRouteProps } from '../../types';
 
 
 function PrivateRoute(props: PrivateRouteProps): JSX.Element {
