+++ src/components/tabs/tabs.tsx
@@ -1,11 +1,11 @@
-import { CITIES } from '../const';
+import { CITIES } from '../../const';
 import React from 'react';
 import classNames from 'classnames';
-import { changeCity } from '../store/actions';
+import { changeCity } from '../../store/actions';
 
 import { Link } from 'react-router-dom';
-import { useAppDispatch } from '../redux-hooks';
-import { TabsProps } from '../types';
+import { useAppDispatch } from '../../redux-hooks';
+import { TabsProps } from '../../types';
 
 function Tabs(props: TabsProps): JSX.Element {
   const {chosenCity} = props;
